- Exclude node modules during build
- Run single Webpack compile

For maximum performance, we should also add the `pg` package to Webpack's `externals`, but doing so seems to break our custom resolve alias. @marvin-kolja Any idea how to do this? Should we just make a custom npm package?

| Project            | Before | After  | Excluding pg |
|--------------------|--------|--------|--------------|
| app-terraform-api  | 39.81s | 2.535s | 1.3575s      |
| app-terraform-user | 3.385s | 2.015s | 1.1775s      |