@marvin-kolja 2022-08-02
> * Exclude node modules during build
> * Run single Webpack compile
> 
> For maximum performance, we should also add the `pg` package to Webpack's `externals`, but doing so seems to break our custom resolve alias. @marvin-kolja Any idea how to do this? Should we just make a custom npm package?

First of all, since this is a public package I would remove all predefined externals and define them using the `nodeExternals` field in the metadata _(or find another way of providing them)_.

Regarding the `pg` package I would need to look into it. It's also unrelated to this package so your idea of creating another package seems good to me.