@marvin-kolja 2022-08-02
Just a follow-up on the last comment:

One option to pass additional config to the package could be through a e.g. `build-lambda-config.js` that could look like this:
```js
const {resolve} = require('path');

module.exports = {
  alias: {
    pg: resolve('./node_modules/@mietz-gmbh/lambda-build-js', 'pg.js')
  },
  externals: [
    'aws-sdk',
    'aws-lambda',
    'aws4',
    'firebase-admin',
    'uuid',
    'lodash',
    'client-oauth2',
    'axios',
    'jsonwebtoken',
    // TODO: Externalize pg without breaking alias
    'pg',
  ]
}

```

Then we can import that file inside this package and use the config like so:

```js
const defaultBuildConfig = {
    alias: {},
    externals: []
}

let buildConfig

try {
    buildConfig = {
        ...defaultBuildConfig,
        ...require(resolve(rootDir, "lambda-build.config.js"))
    };

} catch (error) {
    console.log('Cannot find "lambda-build.config.js"! Using default config.')
    buildConfig = defaultBuildConfig;
}
```