@siphomateke 2022-08-02
@marvin-kolja  Splitting the externals like that would work, but would also mean that we'd need to keep that list in sync in multiple places, right? (e.g. app and user) Ideally, it would be automatically linked to our lambda layer, but at least how it currently is there is only one maintenance point.